add docker volume size to node definition
this makes the size of the docker volume that is attached to each node.
docker uses this volume to fetch images and depending on the size of the
images being used in the cluster the users would like to configure the
size of this volume.

disclaimer: the data structure of nodes/workers is messed up already.
there are differences between aws/azure and kvm whether the nodes can be
configured individually or not. this is a mess and this addition will
not aim to fix this. this will be changed with the work on node pools.