add kubelet volume size to node definition
this makes the size of the kubelet volume that is attached to each node
configurable.

the kubelet uses this volume to store its files and to mount volumes.

disclaimer: the data structure of nodes/workers is messed up already.
there are differences between aws/azure and kvm whether the nodes can be
configured individually or not. this is a mess and this addition will
not aim to fix this. this will be changed with the work on node pools.

built on top of: https://github.com/giantswarm/api-spec/pull/108