deprecated the worker storage configuration
Since the storage configuration is used in KVM only and all use-cases
in which the size was increased are related to docker or kubelet volumes
we decided to remove the option to configure the size of the root
filesystem of the VM. For a transition period the storage size will be
used as the size of the docker volume. But only if the docker volume
itself isn't configured.